properly split registries
