+++ cosmwasm/contracts/crosschain-registry/Cargo.toml
@@ -58,7 +58,7 @@ sha2 = "0.10.6"
 hex = "0.4.3"
 prost = {version = "0.11.2", default-features = false, features = ["prost-derive"]}
 osmosis-std-derive = "0.13.2"
-itertools = "0.10.5"
+registry = { path = "../../packages/registry"}
 
 [dev-dependencies]
 cw-multi-test = "0.16.2"
