+++ cosmwasm/packages/registry/Cargo.toml
@@ -7,8 +7,16 @@ edition = "2021"
 
 [dependencies]
 cosmwasm-schema =  { workspace = true }
-cosmwasm-std =  { workspace = true }
-osmosis-std-derive = "0.13.2"
+cosmwasm-std = { workspace = true }
+cw-storage-plus = { workspace = true }
+serde = { workspace = true }
+thiserror = { workspace = true }
+bech32 = { workspace = true }
 serde-json-wasm = { workspace = true }
+sha2 = "0.10.6"
+hex = "0.4.3"
+schemars = { workspace = true }
+osmosis-std = { workspace = true }
+osmosis-std-derive = "0.13.2"
 prost = {version = "0.11.2", default-features = false, features = ["prost-derive"]}
-crosschain-registry = { path = "../../contracts/crosschain-registry", features = ["imported"]}
+itertools = "0.10.5"
