+++ cosmwasm/packages/registry/src/lib.rs
@@ -4,4 +4,5 @@ mod registry;
 pub use error::RegistryError;
 pub use registry::Registry;
 
-pub mod ibc;
+pub mod msg;
+pub mod proto;
